requirements.txt
fastapi==0.110.1
uvicorn==0.29.0
python-dotenv==1.0.1
llama-index==0.10.28
llama-cloud-services==0.1.3
nest_asyncio==1.6.0
ollama==0.1.7
